django_server/requirements.txt
Django==4.1.13
django-cors-headers==4.3.0
djangorestframework==3.14.0
djongo==1.3.6
openai==1.10.0
Pillow==10.1.0
pymongo==3.12.3
python-dotenv==1.0.0
PyJWT==2.8.0
requests==2.31.0